Terminado lista de productos

+++ Dominio/Articulo.cs
@@ -15,19 +15,20 @@ namespace Dominio
         [DisplayName("Código")]
         public string Codigo { get; set; }
 
-        public string Nombre { get; set;}
+        public string Nombre { get; set; }
 
         [DisplayName("Descripción")]
-        public string Descripcion { get; set;}
+        public string Descripcion { get; set; }
 
-        public Marca Marca { get; set;}
+        public Marca Marca { get; set; }
 
         [DisplayName("Categoría")]
-        public Categoria Categoria { get; set;}
+        public Categoria Categoria { get; set; }
 
-        public string ImagenUrl { get; set;}    
+        public string ImagenUrl { get; set; }
 
-        public decimal Precio { get; set;}
+        public int Precio { get; set; }
+
 
 
     }
